backend/requirements.vm.txt
numpy==2.4.4
Pillow==12.2.0
google-cloud-storage==2.19.0
PyJWT[crypto]>=2.10,<3
python-dotenv==1.2.2
safetensors==0.7.0
transformers==5.5.4
